% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-inference.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Chi-Squared Test}
\usage{
chi_squared_test(data, var1 = NULL, var2 = NULL)
}
\arguments{
\item{data}{A data frame OR a contingency table (matrix)}

\item{var1}{Name of first categorical variable (if data is data frame)}

\item{var2}{Name of second categorical variable (if data is data frame)}
}
\value{
Chi-squared test result
}
\description{
Test independence between two categorical variables.
}
\examples{
\dontrun{
chi_squared_test(mtcars, "am", "cyl")
}
}
