% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-regression.R
\name{compare_models}
\alias{compare_models}
\title{Compare Models}
\usage{
compare_models(model1, model2)
}
\arguments{
\item{model1}{First fitted lm object}

\item{model2}{Second fitted lm object}
}
\value{
ANOVA comparison result
}
\description{
Compare two fitted models using ANOVA F-test.
}
\examples{
\dontrun{
model1 <- fit_linear_model(mtcars, "mpg ~ wt")
model2 <- fit_linear_model(mtcars, "mpg ~ wt + hp")
compare_models(model1, model2)
}
}
