% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-visualization.R
\name{create_comparison_plot}
\alias{create_comparison_plot}
\title{Create Publication-Quality Comparison Plot}
\usage{
create_comparison_plot(data, response, groups)
}
\arguments{
\item{data}{A data frame}

\item{response}{Name of response variable}

\item{groups}{Character vector of grouping variables}
}
\value{
patchwork combined plot
}
\description{
Multi-panel comparison of groups with overlaid distributions.
}
\examples{
\dontrun{
create_comparison_plot(mtcars, "mpg", "cyl")
}
}
