% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-regression.R
\name{fit_polynomial_model}
\alias{fit_polynomial_model}
\title{Fit Polynomial Regression Model}
\usage{
fit_polynomial_model(data, response, predictor, degree = 2)
}
\arguments{
\item{data}{A data frame}

\item{response}{Name of response variable}

\item{predictor}{Name of predictor variable}

\item{degree}{Polynomial degree (default = 2)}
}
\value{
Fitted lm object with polynomial terms
}
\description{
Fit polynomial regression to capture non-linearity.
}
\examples{
\dontrun{
model <- fit_polynomial_model(mtcars, "mpg", "wt", degree = 2)
}
}
