% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{handle_outliers}
\alias{handle_outliers}
\title{Remove or Flag Outliers}
\usage{
handle_outliers(x, method = "flag")
}
\arguments{
\item{x}{A numeric vector}

\item{method}{"remove" or "flag"}
}
\value{
Vector (if method="remove") or data frame with outlier flags
}
\description{
Identify outliers using IQR method (Q1 - 1.5\emph{IQR, Q3 + 1.5}IQR)
}
