% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-inference.R
\name{kruskal_wallis_test}
\alias{kruskal_wallis_test}
\title{Kruskal-Wallis Test (Non-parametric ANOVA)}
\usage{
kruskal_wallis_test(data, response, group)
}
\arguments{
\item{data}{A data frame}

\item{response}{Name of numeric response variable}

\item{group}{Name of grouping variable}
}
\value{
Kruskal-Wallis test result
}
\description{
Test for differences across groups when normality assumption is violated.
}
\examples{
\dontrun{
kruskal_wallis_test(mtcars, response = "mpg", group = "cyl")
}
}
