% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-inference.R
\name{perform_anova}
\alias{perform_anova}
\title{Perform ANOVA}
\usage{
perform_anova(data, response, group)
}
\arguments{
\item{data}{A data frame}

\item{response}{Name of numeric response variable}

\item{group}{Name of grouping variable}
}
\value{
ANOVA object and summary
}
\description{
One-way ANOVA to test for differences across groups.
}
\examples{
\dontrun{
perform_anova(mtcars, response = "mpg", group = "cyl")
}
}
