% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-visualization.R
\name{plot_residual_distribution}
\alias{plot_residual_distribution}
\title{Create Residual Histogram}
\usage{
plot_residual_distribution(model)
}
\arguments{
\item{model}{A fitted lm object}
}
\value{
ggplot object
}
\description{
Visualize distribution of residuals with normality reference.
}
\examples{
\dontrun{
model <- fit_linear_model(mtcars, "mpg ~ wt + hp")
plot_residual_distribution(model)
}
}
