% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{pretty_summary}
\alias{pretty_summary}
\title{Create a Summary Report}
\usage{
pretty_summary(object)
}
\arguments{
\item{object}{Statistical object (lm, aov, etc.)}
}
\value{
Character vector with formatted summary
}
\description{
Generate a human-readable summary of analysis results
}
