% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-stubs.R
\name{transform_data}
\alias{transform_data}
\title{Transform data (light-weight wrapper)}
\usage{
transform_data(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Additional arguments passed to dplyr::mutate (optional)}
}
\value{
A tibble-like data frame
}
\description{
Transform data (light-weight wrapper)
}
