% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{transform_power}
\alias{transform_power}
\title{Apply Power Transformations}
\usage{
transform_power(x, lambda = 0)
}
\arguments{
\item{x}{A numeric vector (must be > 0)}

\item{lambda}{Power to apply; common values: -1 (reciprocal), -0.5, 0 (log), 0.5 (sqrt), 1 (none)}
}
\value{
Transformed vector
}
\description{
Transform a variable using Box-Cox family or common alternatives
}
