% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-inference.R
\name{tukey_post_hoc}
\alias{tukey_post_hoc}
\title{Tukey Post-Hoc Test}
\usage{
tukey_post_hoc(anova_result)
}
\arguments{
\item{anova_result}{ANOVA object from aov()}
}
\value{
Data frame with pairwise comparisons
}
\description{
Pairwise comparisons following ANOVA using Tukey HSD method.
}
\examples{
\dontrun{
anova_result <- perform_anova(mtcars, "mpg", "cyl")
tukey_post_hoc(anova_result)
}
}
