% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_data}
\alias{validate_data}
\title{Validate Data Input}
\usage{
validate_data(data, vars = NULL)
}
\arguments{
\item{data}{A data frame}

\item{vars}{Character vector of variable names to check}
}
\value{
Invisible TRUE or error message
}
\description{
Check that data meets basic requirements for analysis.
}
\keyword{internal}
